apps/api/pyproject.toml
[project]
name = "careeros-api"
version = "0.1.0"
description = "CareerOS - AI Career Intelligence Platform API"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "starlette>=0.41.0",

    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",

    # Validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "authlib>=1.3.0",

    # HTTP client
    "httpx>=0.27.0",

    # Task queue
    "celery[redis]>=5.4.0",
    "redis>=5.0.0",

    # AI / LLM
    "openai>=1.0.0",
    "tiktoken>=0.7.0",

    # Document processing
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
    "chardet>=5.0.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
